Fixes #50
* Some services in models folder have been refactored to services folder.
* Fixes in Gemfile and gemspec to adapt the current source to the last version of Decidim.
