+++ Gemfile
@@ -4,20 +4,19 @@ source 'https://rubygems.org'
 
 ruby '2.4.2'
 
-gem 'decidim', '>= 0.6.0', '<= 0.8.0'
-
 # Declare your gem's dependencies in decidim-initiatives.gemspec.
 # Bundler will treat runtime dependencies like base dependencies, and
 # development dependencies will be added by default to the :development group.
 gemspec
 
+gem 'bootsnap', require: false
+gem 'decidim', '>= 0.7.0', '<= 0.8.0'
+
+
 # Declare any dependencies that are still in development here instead of in
 # your gemspec. These might include edge Rails or gems from your path or
 # Git. Remember to move these dependencies to your gemspec before releasing
 # your gem to rubygems.org.
-
-# To use a debugger
-gem 'rails', '~> 5.1'
 
 group :test, :development do
   gem 'byebug'
