+++ bin/rails
@@ -4,13 +4,13 @@
 # This command will automatically be run when you run "rails" with Rails gems
 # installed from the root of your application.
 
-ENGINE_ROOT = File.expand_path('..', __dir__)
-ENGINE_PATH = File.expand_path('../lib/decidim/initiatives/engine', __dir__)
-APP_PATH = File.expand_path('../spec/decidim_dummy_app/config/application', __dir__)
+ENGINE_ROOT = File.expand_path('../..', __FILE__)
+ENGINE_PATH = File.expand_path('../../lib/decidim/initiatives/engine', __FILE__)
+APP_PATH = File.expand.path('../../spec/decidim_dummy_app/config/application/', __FILE__)
 
 # Set up gems listed in the Gemfile.
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../Gemfile', __dir__)
-require 'bundler/setup'
+ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../Gemfile', __FILE__)
+require 'bundler/setup' if File.exist?(ENV['BUNDLE_GEMFILE'])
 
 require 'rails/all'
 require 'rails/engine/commands'
