+++ decidim-initiatives.gemspec
@@ -21,6 +21,7 @@ Gem::Specification.new do |s|
   s.add_dependency 'decidim-core', '>= 0.7.4', '<= 0.8.0'
   s.add_dependency 'decidim-comments'
   s.add_dependency 'decidim-admin'
+  s.add_dependency 'rails', '~> 5.1.4'
   s.add_dependency 'rectify'
   s.add_dependency 'kaminari', '~> 1.0.1'
   s.add_dependency 'truncato'
