+++ lib/tasks/decidim_initiatives.rake
@@ -27,7 +27,7 @@ namespace :decidim_initiatives do
       .where(second_progress_notification_at: nil).find_each do |initiative|
 
       if initiative.percentage >= Decidim::Initiatives.second_notification_percentage
-        notifier = Decidim::InitiativeProgressNotifier.new(initiative: initiative)
+        notifier = Decidim::Initiatives::ProgressNotifier.new(initiative: initiative)
         notifier.notify
 
         initiative.second_progress_notification_at = DateTime.now
@@ -40,7 +40,7 @@ namespace :decidim_initiatives do
       .where(first_progress_notification_at: nil).find_each do |initiative|
 
       if initiative.percentage >= Decidim::Initiatives.first_notification_percentage
-        notifier = Decidim::InitiativeProgressNotifier.new(initiative: initiative)
+        notifier = Decidim::Initiatives::ProgressNotifier.new(initiative: initiative)
         notifier.notify
 
         initiative.first_progress_notification_at = DateTime.now
